service_notice/websocket_sender/requirements.txt
aio-pika==9.0.4
backoff==2.2.1
pydantic==1.10.2
PyJWT==2.6.0
websockets==10.4
